requirements.txt
Django==5.0.1

django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djoser==2.2.2
social-auth-app-django==5.4.0

django-ckeditor==6.7.0

pytz==2023.3.post1

Pillow==10.2.0
mutagen==1.47.0

psycopg2==2.9.9

argon2-cffi==23.1.0
django-environ==0.11.2

stripe==2.58.0
braintree==4.26.0

pyJWT==2.8.0

boto3==1.34.16
django-storages==1.14.2 

markdown==3.5.2
gunicorn==21.2.0
whitenoise==6.6.0

requests==2.31.0
braintree==4.26.0

stripe==2.58.0